feat: add home-presentation module
This commit introduces the `home-presentation` module, which is a Kotlin Multiplatform library targeting Android and iOS. It includes configuration for Android compilation, iOS XCFramework generation, and source set dependencies. Additionally, it updates the `settings.shared.gradle.kts` file to include the new module in the project structure, and adds a `.gitignore` file.